Add auth check to proxy and integration tests
Add stateless JWT verification to the Next.js proxy (proxy.ts).
Public paths (auth API routes, deny page) bypass the auth check.
Unauthenticated requests redirect to sign-in. Next.js 16 uses
proxy.ts instead of middleware.ts.

Integration tests cover account upsert, session lifecycle,
token_version invalidation, and standard customer access check.

Closes #22
Part of #25
